requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
base58==2.1.1
bitarray==2.7.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.2
cytoolz==0.12.1
Deprecated==1.2.13
Django==4.1.7
django-filter==22.1
django-oauth-toolkit==2.2.0
django-tinymce==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
jwcrypto==1.4.2
lru-dict==1.1.8
Markdown==3.4.1
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
oauthlib==3.2.2
parsimonious==0.8.1
Pillow==9.4.0
protobuf==3.19.5
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
rlp==2.0.1
six==1.16.0
sqlparse==0.4.3
toolz==0.12.0
urllib3==1.26.14
varint==1.0.2
web3==5.31.3
websockets==9.1
wrapt==1.15.0
yarl==1.8.2
